Added COMMTEST packet types, added List Mode data record loop
- Added the commtest packet types for stream start and stopping (will be
needed for testing)
- Added public functions to the HardwareInterface to start and stop the
list mode streaming. This list datapacket is forwarded to a filestream.
The SYNC and CHECKSUM bytes are stripped from the packet, but PID and
LEN are still included. This can be used later to finde FIFO overflows
or add other type of data to the stream by adding custom PIDs